requirements.txt
absl-py==1.0.0
appdirs==1.4.3
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
blinker==1.4
cachetools==4.2.4
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloudpickle==1.6.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
deap==1.3.1
decorator==4.4.2
distlib==0.3.0
distro==1.7.0
dm-tree==0.1.7
entrypoints==0.3
filelock==3.0.12
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gym==0.21.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.8.2
incremental==16.10.1
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jeepney==0.8.0
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==1.1.0
matplotlib==3.3.3
mccabe==0.6.1
more-itertools==4.2.0
mpi4py==3.1.3
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.5
parso==0.7.1
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybullet==3.2.4
pycparser==2.20
pyglet==1.5.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
pytest==6.1.2
python-dateutil==2.8.1
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
seaborn==0.11.1
SecretStorage==3.3.2
simplejson==3.16.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-probability==0.16.0
termcolor==1.1.0
testresources==2.0.1
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
torch==1.12.1
tornado==6.1
traitlets==5.0.5
typing-extensions==4.0.1
urllib3==1.26.9
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
-e .
